requirements.txt
google-genai>=1.51.0
openai>=1.0.0
yfinance>=1.0
pandas>=2.0
numpy>=1.24
python-dotenv>=1.0.0
ta>=0.11.0
pytest>=8.0.0
